+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 
+## 0.0.7 (2022-11-08)
+- Add Action to publish on PyPI [#1](https://github.com/lucasrcezimbra/agilize/issues/1)
+- Fix Prolabore.from_data to cast values to Decimal
+- Improve Competence adding `previous`, `next`, `last_date`, `from_date`, `first_date`
+- Update dev requirements
+
+
 ## 0.0.6 (2022-10-29)
 - Enable NFSE upload - Agilize.upload_nfse and Client.upload_nfse
 - Update dev requirements